+++ changelog.md
@@ -1,3 +1,21 @@
+# v.1.0003
+
+### pack info
+- level mods can now use pack info scenes to display information about the mod
+- for more info, see the *pack info* section in the modarchy text tutorial
+
+### pause fixes
+- use datahelper to store volume
+- preload volume and window text
+- add keybinds for toggling pause menu
+- turning on fullscreen now consistently works
+
+### miscellaneous
+- `on_floor` is now a public variable
+- mods will only be unloaded if the `_mods` folder exists
+
+---
+
 # v.1.0002
 
 ### pause menu